requirements.txt
# FOMC Project Dependencies

# Core dependencies
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# API dependencies
requests==2.31.0

# Data processing
pandas==2.1.3
numpy==1.26.2

# Web framework
fastapi==0.104.1
uvicorn==0.24.0

# Visualization (to be used later)
matplotlib==3.8.2
seaborn==0.13.0

# Development dependencies
pytest==7.4.3
black==23.11.0
flake8==6.1.0